// backend/src/generated/zod/inputTypeSchemas/MetricMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const MetricMinOrderByAggregateInputSchema: z.ZodType<Prisma.MetricMinOrderByAggregateInput> = z.strictObject({
  id: z.lazy(() => SortOrderSchema).optional(),
  tenantId: z.lazy(() => SortOrderSchema).optional(),
  serviceName: z.lazy(() => SortOrderSchema).optional(),
  metricName: z.lazy(() => SortOrderSchema).optional(),
  timestamp: z.lazy(() => SortOrderSchema).optional(),
  value: z.lazy(() => SortOrderSchema).optional(),
});

export default MetricMinOrderByAggregateInputSchema;